refactor: split PySocket into net::socket
